docs/requirements.txt
mkdocs-techdocs-diamond
mkdocstrings[python]
ruamel.yaml